test: fix integer overflow in IconAssetTests bounds check
offset + length used int arithmetic and could wrap negative on hostile
values, falling through to the exception it was meant to replace. Guard
against negative offset/length and widen the sum to long before comparing
against the file length.

+++ AudioLeash.Tests/IconAssetTests.cs
@@ -34,7 +34,7 @@ public class IconAssetTests
             int length = BitConverter.ToInt32(ico, entry + 8);
             int offset = BitConverter.ToInt32(ico, entry + 12);
 
-            if (offset < 0 || length < 0 || offset + length > ico.Length)
+            if (offset < 0 || length < 0 || (long)offset + length > ico.Length)
             {
                 throw new InvalidOperationException(
                     $"icon.ico frame {i} has an out-of-range payload (offset={offset}, " +
